requirements.txt
numpy
scipy
opencv-python
PyYAML
matplotlib
numba
pytest
torch
timm
